pet support and auto camo
Ok so pet update today, yay!

Here is how it works:
We have 2 new callbacks added to the swf and one new Handler:

petCall-
takes a positive integer as it's only argument, that integer will
be the ID of the command you want to execute, the commands are:
0 - Activate
1 - Deactivate
4 - repair

PetModule-
Takes a positive integer as the first argument and a (usually empty) string as the second.
The first argument is the module ID and the seconds is usually an empty string.

the IDs i know so far are the following:
KAMIKAZE = 10;
RESOURCE_LOCATOR = 7;
ENEMY_LOCATOR = 6;
AUTO_LOOT = 4;
REPAIR_PET = 9;
TRADE_POD = 8;
COMBO_GUARD = 12;
AUTO_RESOURCE_COLLECTION = 5;
FRIENDLY_SACRIFICE = 14;
COMBO_SHIP_REPAIR = 11;

PetUpdateFuel handler-
Called everytime the pet loses fuel.

Well, i wish to explain the javascript side of things but it's just a big mess that somehow works.

bugs:
-Module doesn't change in pet interface.
-Bot doesn't know when pet is deactivated (by the user, for example).
This can be solved by checking for the pet ID on the ShipRemovedHandler. Carefull because ShipRemovedHandler gets called before ShipDestroyed handler,
 where we check if the pet was destroyed.