Testing features and removing files
Added a changeMode option that changes to flying mode while looking for npc or boxes. Note: it doesn't change while collecting, only while there are no boxes or npcs nearby.

Added a sleep() function in case needed in the future. I wanted to add it a while ago but forgo the reason why.

Removed 2 useless handlers.

BOT IS STILL OFFLINE!